# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_stuartshay_gcp-automation-api_internal_models.AdvancedBucketRequest:
    properties:
      kms_key_name:
        description: Phase 1 Advanced Options - Security & Compliance
        example: projects/velvety-byway-327718/locations/us-central1/keyRings/bucket-encryption/cryptoKeys/bucket-key
        type: string
      labels:
        additionalProperties:
          type: string
        example:
          '"cost-center"': '"engineering"}'
          '"team"': '"platform"'
          '{"environment"': '"production"'
        type: object
      location:
        example: us-central1
        type: string
      name:
        example: my-enterprise-bucket-2024
        type: string
      public_access_prevention:
        enum:
        - inherited
        - enforced
        - unspecified
        example: enforced
        type: string
      retention_policy:
        $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.RetentionPolicy'
      storage_class:
        enum:
        - STANDARD
        - NEARLINE
        - COLDLINE
        - ARCHIVE
        example: STANDARD
        type: string
      uniform_bucket_level_access:
        example: true
        type: boolean
      versioning:
        example: true
        type: boolean
    required:
    - location
    - name
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.AdvancedFolderRequest:
    properties:
      display_name:
        example: Production - North America Region
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        example: "987654321098"
        type: string
      parent_type:
        enum:
        - organization
        - folder
        example: folder
        type: string
    required:
    - display_name
    - parent_id
    - parent_type
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.AdvancedProjectRequest:
    properties:
      display_name:
        example: Enterprise Application - Production
        maxLength: 100
        minLength: 1
        type: string
      labels:
        additionalProperties:
          type: string
        example:
          '"compliance"': '"sox"}'
          '"cost-center"': '"engineering"'
          '"team"': '"backend"'
          '{"environment"': '"production"'
        type: object
      parent_id:
        example: "123456789012"
        type: string
      parent_type:
        enum:
        - organization
        - folder
        example: organization
        type: string
      project_id:
        example: enterprise-app-prod-2024
        type: string
    required:
    - display_name
    - project_id
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.BasicBucketRequest:
    properties:
      location:
        example: us-central1
        type: string
      name:
        example: my-simple-storage-bucket
        type: string
      storage_class:
        enum:
        - STANDARD
        - NEARLINE
        - COLDLINE
        - ARCHIVE
        example: STANDARD
        type: string
    required:
    - location
    - name
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.BasicFolderRequest:
    properties:
      display_name:
        example: Development Environment
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        example: "123456789012"
        type: string
      parent_type:
        enum:
        - organization
        - folder
        example: organization
        type: string
    required:
    - display_name
    - parent_id
    - parent_type
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.BasicProjectRequest:
    properties:
      display_name:
        example: My Simple Project
        maxLength: 100
        minLength: 1
        type: string
      project_id:
        example: my-simple-project-2024
        type: string
    required:
    - display_name
    - project_id
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.BucketRequest:
    properties:
      kms_key_name:
        description: Phase 1 Advanced Options - Security & Compliance
        example: projects/my-project/locations/us-central1/keyRings/my-keyring/cryptoKeys/my-key
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      location:
        example: us-central1
        type: string
      name:
        example: my-project-storage-bucket
        type: string
      public_access_prevention:
        enum:
        - inherited
        - enforced
        - unspecified
        example: enforced
        type: string
      retention_policy:
        $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.RetentionPolicy'
      storage_class:
        enum:
        - STANDARD
        - NEARLINE
        - COLDLINE
        - ARCHIVE
        example: STANDARD
        type: string
      uniform_bucket_level_access:
        example: true
        type: boolean
      versioning:
        example: true
        type: boolean
    required:
    - location
    - name
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.BucketResponse:
    properties:
      create_time:
        type: string
      kms_key_name:
        description: Phase 1 Advanced Options - Security & Compliance
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      location:
        type: string
      name:
        type: string
      public_access_prevention:
        type: string
      retention_policy:
        $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.RetentionPolicy'
      self_link:
        type: string
      storage_class:
        type: string
      uniform_bucket_level_access:
        type: boolean
      update_time:
        type: string
      versioning:
        type: boolean
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.FolderRequest:
    properties:
      display_name:
        example: Development Environment
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        example: "123456789012"
        type: string
      parent_type:
        description: '"organization" or "folder"'
        enum:
        - organization
        - folder
        example: organization
        type: string
    required:
    - display_name
    - parent_id
    - parent_type
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.FolderResponse:
    properties:
      create_time:
        type: string
      display_name:
        type: string
      name:
        type: string
      parent_id:
        type: string
      parent_type:
        type: string
      state:
        type: string
      update_time:
        type: string
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.ProjectRequest:
    properties:
      display_name:
        example: My Development Project
        maxLength: 100
        minLength: 1
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      parent_id:
        example: "123456789012"
        type: string
      parent_type:
        description: '"organization" or "folder"'
        enum:
        - organization
        - folder
        example: organization
        type: string
      project_id:
        example: my-dev-project-2024
        type: string
    required:
    - display_name
    - project_id
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.ProjectResponse:
    properties:
      create_time:
        type: string
      display_name:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      parent_id:
        type: string
      parent_type:
        type: string
      project_id:
        type: string
      project_number:
        type: integer
      state:
        type: string
      update_time:
        type: string
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.RetentionPolicy:
    properties:
      is_locked:
        example: false
        type: boolean
      retention_period_seconds:
        description: 1 second to 100 years
        example: 86400
        maximum: 3155760000
        minimum: 1
        type: integer
    type: object
  github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for automating Google Cloud Platform resource management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GCP Automation API
  version: "1.0"
paths:
  /buckets:
    post:
      consumes:
      - application/json
      description: |-
        Create a new Google Cloud Storage bucket with the specified parameters including advanced options for KMS encryption, retention policies, and access controls

        ## Example Usage:
        ### Basic Example (models.BasicBucketRequest):
        Simple bucket creation with minimal required fields
        ### Advanced Example (models.AdvancedBucketRequest):
        Enterprise bucket with security features, labels, and compliance settings
      parameters:
      - description: Bucket creation request
        in: body
        name: bucket
        required: true
        schema:
          $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.BucketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.BucketResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new Cloud Storage bucket
      tags:
      - Buckets
  /buckets/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a Google Cloud Storage bucket by its bucket name
      parameters:
      - description: Bucket name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a Cloud Storage bucket
      tags:
      - Buckets
    get:
      consumes:
      - application/json
      description: Retrieve details of a Google Cloud Storage bucket by its bucket
        name
      parameters:
      - description: Bucket name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.BucketResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a Cloud Storage bucket by name
      tags:
      - Buckets
  /examples/buckets:
    get:
      consumes:
      - application/json
      description: Provides both basic and advanced example request bodies for creating
        GCS buckets
      produces:
      - application/json
      responses:
        "200":
          description: Advanced bucket creation example
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.AdvancedBucketRequest'
              type: object
      summary: Get example request bodies for bucket creation
      tags:
      - Examples
  /examples/folders:
    get:
      consumes:
      - application/json
      description: Provides both basic and advanced example request bodies for creating
        GCP folders
      produces:
      - application/json
      responses:
        "200":
          description: Advanced folder creation example
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.AdvancedFolderRequest'
              type: object
      summary: Get example request bodies for folder creation
      tags:
      - Examples
  /examples/projects:
    get:
      consumes:
      - application/json
      description: Provides both basic and advanced example request bodies for creating
        GCP projects
      produces:
      - application/json
      responses:
        "200":
          description: Advanced project creation example
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.AdvancedProjectRequest'
              type: object
      summary: Get example request bodies for project creation
      tags:
      - Examples
  /folders:
    post:
      consumes:
      - application/json
      description: |-
        Create a new Google Cloud Platform folder with the specified parameters

        ## Example Usage:
        ### Basic Example (models.BasicFolderRequest):
        Simple folder under organization root
        ### Advanced Example (models.AdvancedFolderRequest):
        Nested folder structure for complex hierarchies
      parameters:
      - description: Folder creation request
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.FolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.FolderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new GCP folder
      tags:
      - Folders
  /folders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Google Cloud Platform folder by its folder ID
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a GCP folder
      tags:
      - Folders
    get:
      consumes:
      - application/json
      description: Retrieve details of a Google Cloud Platform folder by its folder
        ID
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.FolderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a GCP folder by ID
      tags:
      - Folders
  /projects:
    post:
      consumes:
      - application/json
      description: |-
        Create a new Google Cloud Platform project with the specified parameters

        ## Example Usage:
        ### Basic Example (models.BasicProjectRequest):
        Simple project creation with just ID and display name
        ### Advanced Example (models.AdvancedProjectRequest):
        Enterprise project with organization hierarchy, labels, and governance
      parameters:
      - description: Project creation request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new GCP project
      tags:
      - Projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Google Cloud Platform project by its project ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a GCP project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Retrieve details of a Google Cloud Platform project by its project
        ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stuartshay_gcp-automation-api_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a GCP project by ID
      tags:
      - Projects
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"